requirements.txt
ultralytics==8.0
numpy
opencv-python
pandas
torch
Pillow==9.5
albumentations==1.0.3
pyyaml
